+++ deno_dist/core/menu/map/index.ts
@@ -1,4 +1,5 @@
 export { HashMap as Hashed } from '../../../hashed/map/index.ts';
 export { SortedMap as Sorted } from '../../../sorted/map/index.ts';
+export { ProximityMap as Proximity } from '../../../proximity/map/index.ts';
 
 export * as Ordered from './ordered.ts';