+++ deno_dist/proximity/common/index.ts
@@ -1,0 +1,2 @@
+export * from './distanceFunction.ts';
+export * from './keyMatching.ts';