+++ deno_dist/actor/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.11.7](https://github.com/rimbu-org/rimbu/compare/@rimbu/actor@0.11.6...@rimbu/actor@0.11.7) (2023-05-21)
+
+**Note:** Version bump only for package @rimbu/actor
+
 ## [0.11.6](https://github.com/rimbu-org/rimbu/compare/@rimbu/actor@0.11.5...@rimbu/actor@0.11.6) (2023-04-30)
 
 **Note:** Version bump only for package @rimbu/actor