+++ deno_dist/core/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.11.6](https://github.com/rimbu-org/rimbu/compare/@rimbu/core@0.11.5...@rimbu/core@0.11.6) (2023-05-21)
+
+**Note:** Version bump only for package @rimbu/core
+
 ## [0.11.5](https://github.com/rimbu-org/rimbu/compare/@rimbu/core@0.11.4...@rimbu/core@0.11.5) (2023-04-30)
 
 **Note:** Version bump only for package @rimbu/core