+++ deno_dist/list/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.2](https://github.com/rimbu-org/rimbu/compare/@rimbu/list@0.12.1...@rimbu/list@0.12.2) (2023-05-21)
+
+**Note:** Version bump only for package @rimbu/list
+
 ## [0.12.1](https://github.com/rimbu-org/rimbu/compare/@rimbu/list@0.12.0...@rimbu/list@0.12.1) (2023-04-30)
 
 **Note:** Version bump only for package @rimbu/list