+++ deno_dist/ordered/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.6](https://github.com/rimbu-org/rimbu/compare/@rimbu/ordered@0.9.5...@rimbu/ordered@0.9.6) (2023-05-21)
+
+**Note:** Version bump only for package @rimbu/ordered
+
 ## [0.9.5](https://github.com/rimbu-org/rimbu/compare/@rimbu/ordered@0.9.4...@rimbu/ordered@0.9.5) (2023-04-30)
 
 **Note:** Version bump only for package @rimbu/ordered